move pollPendingData into sync.go
